requirements.txt
tensorflow-gpu
tensorflow-addons
numpy
scipy
opencv-python
opencv-contrib-python
tqdm
pyyaml
flake8
jupyter
